pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "docx2pdf>=0.1.8",
    "google-genai>=1.27.0",
    "llama-cloud-services>=0.6.51",
    "llama-index-vector-stores-pinecone>=0.6.0",
    "llama-index>=0.12.52",
    "mammoth>=1.9.1",
    "pdfkit>=1.0.0",
    "pinecone>=7.3.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.3",
    "requests>=2.32.4",
    "weasyprint>=66.0",
    "langchain-google-genai>=2.1.8",
    "langchain-community>=0.3.27",
    "unstructured[md]>=0.18.11",
    "llama-index-embeddings-google-genai>=0.2.1",
    "llama-index-llms-google-genai>=0.2.5",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "spire-doc-free>=12.12.0",
    "chromadb>=1.0.15",
    "extract-msg>=0.41.0",
    "email-validator>=2.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "opencv-python>=4.12.0.88",
    "langgraph>=0.6.1",
    "pytest>=8.4.1",
    "llama-index-vector-stores-chroma>=0.4.2",
]
